@@ -1,6 +1,6 @@
 {
   "name": "sol2uml",
-  "version": "1.1.25",
+  "version": "1.1.30",
   "description": "Unified Modeling Language (UML) class diagram generator for Solidity contracts",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
@@ -19,10 +19,10 @@
   "repository": "github:naddison36/sol2uml",
   "license": "MIT",
   "dependencies": {
-    "@solidity-parser/parser": "^0.10.2",
+    "@solidity-parser/parser": "^0.13.2",
     "axios": "^0.21.1",
-    "commander": "^3.0.2",
-    "debug": "^4.3.1",
+    "commander": "^8.1.0",
+    "debug": "^4.3.2",
     "js-graph-algorithms": "^1.0.18",
     "klaw": "^3.0.0",
     "svg-to-png": "^4.0.0",
@@ -30,15 +30,15 @@
     "viz.js": "^1.8.2"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.19",
-    "@types/klaw": "^3.0.1",
-    "@types/verror": "^1.10.4",
-    "jest": "^26.6.3",
+    "@types/jest": "^26.0.24",
+    "@types/klaw": "^3.0.2",
+    "@types/verror": "^1.10.5",
+    "jest": "^27.0.6",
     "openzeppelin-solidity": "2.5.1",
-    "prettier": "^2.2.1",
-    "ts-jest": "^26.4.4",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.1.3"
+    "prettier": "^2.3.2",
+    "ts-jest": "^27.0.4",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   },
   "files": [
     "lib/*.js",
