@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.writePng = exports.writeSVG = exports.writeDot = exports.convertDot2Svg = exports.addAssociationsToDot = exports.convertUmlClasses2Dot = exports.convertUmlClassesToSvg = exports.generateFilesFromUmlClasses = void 0;
+exports.writePng = exports.writeSVG = exports.writeDot = exports.writeSolidity = exports.convertDot2Svg = exports.addAssociationsToDot = exports.convertUmlClasses2Dot = exports.convertUmlClassesToSvg = exports.generateFilesFromUmlClasses = void 0;
 const fs_1 = require("fs");
 const path = require('path');
 const svg_to_png = require('svg-to-png');
@@ -127,7 +127,11 @@ function addAssociationToDot(sourceUmlClass, targetUmlClass, association, classO
         (sourceUmlClass.stereotype === umlClass_1.ClassStereotype.Library ||
             targetUmlClass.stereotype === umlClass_1.ClassStereotype.Library)) ||
         (classOptions.hideInterfaces &&
-            targetUmlClass.stereotype === umlClass_1.ClassStereotype.Interface)) {
+            targetUmlClass.stereotype === umlClass_1.ClassStereotype.Interface) ||
+        (classOptions.hideStructs &&
+            targetUmlClass.stereotype === umlClass_1.ClassStereotype.Struct) ||
+        (classOptions.hideEnums &&
+            targetUmlClass.stereotype === umlClass_1.ClassStereotype.Enum)) {
         return '';
     }
     let dotString = `\n${sourceUmlClass.id} -> ${targetUmlClass.id} [`;
@@ -159,6 +163,19 @@ function convertDot2Svg(dot) {
     }
 }
 exports.convertDot2Svg = convertDot2Svg;
+function writeSolidity(code, filename = 'solidity') {
+    const outputFile = filename + '.sol';
+    debug(`About to write Solidity to file ${outputFile}`);
+    fs_1.writeFile(outputFile, code, (err) => {
+        if (err) {
+            throw new verror_1.VError(err, `Failed to write Solidity to file ${outputFile}`);
+        }
+        else {
+            console.log(`Solidity written to ${outputFile}`);
+        }
+    });
+}
+exports.writeSolidity = writeSolidity;
 function writeDot(dot, dotFilename = 'classDiagram.dot') {
     debug(`About to write Dot file to ${dotFilename}`);
     fs_1.writeFile(dotFilename, dot, (err) => {
