@@ -4,21 +4,24 @@ exports.dotUmlClass = void 0;
 // Returns a string of the UML Class in Graphviz's dot format
 const umlClass_1 = require("./umlClass");
 const dotUmlClass = (umlClass, options = {}) => {
-    // do not include library or interface classes if hidden
+    // do not include library, interface, struct or enum classes if hidden
     if ((options.hideLibraries &&
         umlClass.stereotype === umlClass_1.ClassStereotype.Library) ||
         (options.hideInterfaces &&
-            umlClass.stereotype === umlClass_1.ClassStereotype.Interface)) {
+            umlClass.stereotype === umlClass_1.ClassStereotype.Interface) ||
+        (options.hideStructs &&
+            umlClass.stereotype === umlClass_1.ClassStereotype.Struct) ||
+        (options.hideEnums && umlClass.stereotype === umlClass_1.ClassStereotype.Enum)) {
         return '';
     }
     let dotString = `\n${umlClass.id} [label="{${dotClassTitle(umlClass)}`;
     // Add attributes
     if (!options.hideAttributes) {
-        dotString += dotAttributeVisibilities(umlClass);
+        dotString += dotAttributeVisibilities(umlClass, options);
     }
     // Add operators
     if (!options.hideOperators) {
-        dotString += dotOperatorVisibilities(umlClass);
+        dotString += dotOperatorVisibilities(umlClass, options);
     }
     dotString += '}"]';
     // Output structs and enums
@@ -43,24 +46,37 @@ const dotClassTitle = (umlClass) => {
         case umlClass_1.ClassStereotype.Library:
             stereoName = 'Library';
             break;
+        case umlClass_1.ClassStereotype.Struct:
+            stereoName = 'Struct';
+            break;
+        case umlClass_1.ClassStereotype.Enum:
+            stereoName = 'Enum';
+            break;
         default:
             // Contract or undefined stereotype will just return the UmlClass name
             return umlClass.name;
     }
     return `\\<\\<${stereoName}\\>\\>\\n${umlClass.name}`;
 };
-const dotAttributeVisibilities = (umlClass) => {
+const dotAttributeVisibilities = (umlClass, options = {}) => {
     let dotString = '| ';
+    // if a struct or enum then no visibility group
+    if (umlClass.stereotype === umlClass_1.ClassStereotype.Struct ||
+        umlClass.stereotype === umlClass_1.ClassStereotype.Enum) {
+        return dotString + dotAttributes(umlClass.attributes, undefined, false);
+    }
     // For each visibility group
     for (const vizGroup of ['Private', 'Internal', 'External', 'Public']) {
         const attributes = [];
         // For each attribute of te UML Class
         for (const attribute of umlClass.attributes) {
-            if (vizGroup === 'Private' &&
+            if (!options.hideInternals &&
+                vizGroup === 'Private' &&
                 attribute.visibility === umlClass_1.Visibility.Private) {
                 attributes.push(attribute);
             }
-            else if (vizGroup === 'Internal' &&
+            else if (!options.hideInternals &&
+                vizGroup === 'Internal' &&
                 attribute.visibility === umlClass_1.Visibility.Internal) {
                 attributes.push(attribute);
             }
@@ -76,33 +92,36 @@ const dotAttributeVisibilities = (umlClass) => {
                 attributes.push(attribute);
             }
         }
-        dotString += dotAttributes(vizGroup, attributes);
+        dotString += dotAttributes(attributes, vizGroup);
     }
     return dotString;
 };
-const dotAttributes = (vizGroup, attributes) => {
+const dotAttributes = (attributes, vizGroup, indent = true) => {
     if (!attributes || attributes.length === 0) {
         return '';
     }
-    let dotString = vizGroup + ':\\l';
+    const indentString = indent ? '\\ \\ \\ ' : '';
+    let dotString = vizGroup ? vizGroup + ':\\l' : '';
     // for each attribute
     attributes.forEach((attribute) => {
-        dotString += `\\ \\ \\ ${attribute.name}: ${attribute.type}\\l`;
+        dotString += `${indentString}${attribute.name}: ${attribute.type}\\l`;
     });
     return dotString;
 };
-const dotOperatorVisibilities = (umlClass) => {
+const dotOperatorVisibilities = (umlClass, options = {}) => {
     let dotString = '| ';
     // For each visibility group
     for (const vizGroup of ['Private', 'Internal', 'External', 'Public']) {
         const operators = [];
         // For each attribute of te UML Class
         for (const operator of umlClass.operators) {
-            if (vizGroup === 'Private' &&
+            if (!options.hideInternals &&
+                vizGroup === 'Private' &&
                 operator.visibility === umlClass_1.Visibility.Private) {
                 operators.push(operator);
             }
-            else if (vizGroup === 'Internal' &&
+            else if (!options.hideInternals &&
+                vizGroup === 'Internal' &&
                 operator.visibility === umlClass_1.Visibility.Internal) {
                 operators.push(operator);
             }
@@ -223,8 +242,9 @@ const dotEnums = (umlClass) => {
         const enumId = umlClass.id + 'enum' + enumCount++;
         dotString += `\n"${enumId}" [label="{\\<\\<enum\\>\\>\\n${enumKey}|`;
         // output each enum value
+        let enumIndex = 0;
         for (const value of umlClass.enums[enumKey]) {
-            dotString += value + '\\l';
+            dotString += value + ': ' + enumIndex++ + '\\l';
         }
         dotString += '}"]';
         // Add the association to the contract the enum was declared in
