@@ -1,6 +1,6 @@
-import { ASTNode } from '@solidity-parser/parser/dist/ast-types';
+import { ASTNode } from '@solidity-parser/parser/dist/src/ast-types';
 import { UmlClass } from './umlClass';
-declare const networks: readonly ["mainnet", "ropsten", "kovan", "rinkeby", "goerli"];
+declare const networks: readonly ["mainnet", "ropsten", "kovan", "rinkeby", "goerli", "polygon", "bsc"];
 declare type Network = typeof networks[number];
 export declare class EtherscanParser {
     protected apikey: string;
@@ -13,6 +13,13 @@ export declare class EtherscanParser {
      * @return Promise with an array of UmlClass objects
      */
     getUmlClasses(contractAddress: string): Promise<UmlClass[]>;
+    /**
+     * Get Solidity code from Etherscan for a contract and merges all files
+     * into one long string of Solidity code.
+     * @param contractAddress Ethereum contract address with a 0x prefix
+     * @return Promise string of Solidity code
+     */
+    getSolidityCode(contractAddress: string): Promise<string>;
     /**
      * Parses Solidity source code into an ASTNode object
      * @param sourceCode Solidity source code
