@@ -8,7 +8,15 @@ const axios_1 = __importDefault(require("axios"));
 const parser_1 = require("@solidity-parser/parser");
 const verror_1 = require("verror");
 const parser_2 = require("./parser");
-const networks = ['mainnet', 'ropsten', 'kovan', 'rinkeby', 'goerli'];
+const networks = [
+    'mainnet',
+    'ropsten',
+    'kovan',
+    'rinkeby',
+    'goerli',
+    'polygon',
+    'bsc',
+];
 class EtherscanParser {
     constructor(apikey = 'ZAD4UI2RCXCQTP38EXS3UY2MPHFU5H9KB1', network = 'mainnet') {
         this.apikey = apikey;
@@ -18,6 +26,14 @@ class EtherscanParser {
         }
         else if (network === 'mainnet') {
             this.url = 'https://api.etherscan.io/api';
+        }
+        else if (network === 'polygon') {
+            this.url = 'https://api.polygonscan.com/api';
+            this.apikey = 'AMHGNTV5A7XYGX2M781JB3RC1DZFVRWQEB';
+        }
+        else if (network === 'bsc') {
+            this.url = 'https://api.bscscan.com/api';
+            this.apikey = 'APYH49FXVY9UA3KTDI6F4WP3KPIC86NITN';
         }
         else {
             this.url = `https://api-${network}.etherscan.io/api`;
@@ -37,6 +53,20 @@ class EtherscanParser {
             umlClasses = umlClasses.concat(umlClass);
         }
         return umlClasses;
+    }
+    /**
+     * Get Solidity code from Etherscan for a contract and merges all files
+     * into one long string of Solidity code.
+     * @param contractAddress Ethereum contract address with a 0x prefix
+     * @return Promise string of Solidity code
+     */
+    async getSolidityCode(contractAddress) {
+        const sourceFiles = await this.getSourceCode(contractAddress);
+        let solidityCode = '';
+        sourceFiles.forEach((sourceFile) => {
+            solidityCode += sourceFile.code;
+        });
+        return solidityCode;
     }
     /**
      * Parses Solidity source code into an ASTNode object
