@@ -22,6 +22,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.convertNodeToUmlClass = void 0;
 const path = __importStar(require("path"));
 const umlClass_1 = require("./umlClass");
+const typeGuards_1 = require("./typeGuards");
 const debug = require('debug')('sol2uml');
 function convertNodeToUmlClass(node, relativePath, filesystem = false) {
     let umlClasses = [];
@@ -38,6 +39,32 @@ function convertNodeToUmlClass(node, relativePath, filesystem = false) {
                     relativePath,
                 });
                 umlClass = parseContractDefinition(umlClass, childNode);
+                umlClasses.push(umlClass);
+            }
+            else if (childNode.type === 'StructDefinition') {
+                debug(`Adding struct ${childNode.name}`);
+                let umlClass = new umlClass_1.UmlClass({
+                    name: childNode.name,
+                    stereotype: umlClass_1.ClassStereotype.Struct,
+                    absolutePath: filesystem
+                        ? path.resolve(relativePath) // resolve the absolute path
+                        : relativePath,
+                    relativePath,
+                });
+                umlClass = parseStructDefinition(umlClass, childNode);
+                umlClasses.push(umlClass);
+            }
+            else if (childNode.type === 'EnumDefinition') {
+                debug(`Adding enum ${childNode.name}`);
+                let umlClass = new umlClass_1.UmlClass({
+                    name: childNode.name,
+                    stereotype: umlClass_1.ClassStereotype.Enum,
+                    absolutePath: filesystem
+                        ? path.resolve(relativePath) // resolve the absolute path
+                        : relativePath,
+                    relativePath,
+                });
+                umlClass = parseEnumDefinition(umlClass, childNode);
                 umlClasses.push(umlClass);
             }
             else if (childNode.type === 'ImportDirective') {
@@ -71,6 +98,29 @@ function convertNodeToUmlClass(node, relativePath, filesystem = false) {
     return umlClasses;
 }
 exports.convertNodeToUmlClass = convertNodeToUmlClass;
+function parseStructDefinition(umlClass, node) {
+    node.members.forEach((member) => {
+        umlClass.attributes.push({
+            name: member.name,
+            type: parseTypeName(member.typeName),
+        });
+    });
+    // Recursively parse struct members for associations
+    umlClass = addAssociations(node.members, umlClass);
+    return umlClass;
+}
+function parseEnumDefinition(umlClass, node) {
+    let index = 0;
+    node.members.forEach((member) => {
+        umlClass.attributes.push({
+            name: member.name,
+            type: (index++).toString(),
+        });
+    });
+    // Recursively parse struct members for associations
+    umlClass = addAssociations(node.members, umlClass);
+    return umlClass;
+}
 function parseContractDefinition(umlClass, node) {
     umlClass.stereotype = parseContractKind(node.kind);
     // For each base contract
@@ -84,114 +134,112 @@ function parseContractDefinition(umlClass, node) {
     });
     // For each sub node
     node.subNodes.forEach((subNode) => {
-        switch (subNode.type) {
-            case 'StateVariableDeclaration':
-                subNode.variables.forEach((variable) => {
-                    umlClass.attributes.push({
-                        visibility: parseVisibility(variable.visibility),
-                        name: variable.name,
-                        type: parseTypeName(variable.typeName),
-                    });
+        if (typeGuards_1.isStateVariableDeclaration(subNode)) {
+            subNode.variables.forEach((variable) => {
+                umlClass.attributes.push({
+                    visibility: parseVisibility(variable.visibility),
+                    name: variable.name,
+                    type: parseTypeName(variable.typeName),
                 });
-                // Recursively parse variables for associations
-                umlClass = addAssociations(subNode.variables, umlClass);
-                break;
-            case 'UsingForDeclaration':
-                // Add association to library contract
-                umlClass.addAssociation({
-                    referenceType: umlClass_1.ReferenceType.Memory,
-                    targetUmlClassName: subNode.libraryName,
+            });
+            // Recursively parse variables for associations
+            umlClass = addAssociations(subNode.variables, umlClass);
+        }
+        else if (typeGuards_1.isUsingForDeclaration(subNode)) {
+            // Add association to library contract
+            umlClass.addAssociation({
+                referenceType: umlClass_1.ReferenceType.Memory,
+                targetUmlClassName: subNode.libraryName,
+            });
+        }
+        else if (typeGuards_1.isFunctionDefinition(subNode)) {
+            if (subNode.isConstructor) {
+                umlClass.operators.push({
+                    name: 'constructor',
+                    stereotype: umlClass_1.OperatorStereotype.None,
+                    parameters: parseParameters(subNode.parameters),
                 });
-                break;
-            case 'FunctionDefinition':
-                if (subNode.isConstructor) {
-                    umlClass.operators.push({
-                        name: 'constructor',
-                        stereotype: umlClass_1.OperatorStereotype.None,
-                        parameters: parseParameters(subNode.parameters),
-                    });
-                }
-                // If a fallback function
-                else if (subNode.name === '') {
-                    umlClass.operators.push({
-                        name: '',
-                        stereotype: umlClass_1.OperatorStereotype.Fallback,
-                        parameters: parseParameters(subNode.parameters),
-                        isPayable: parsePayable(subNode.stateMutability),
-                    });
-                }
-                else {
-                    let stereotype = umlClass_1.OperatorStereotype.None;
-                    if (subNode.body === null) {
-                        stereotype = umlClass_1.OperatorStereotype.Abstract;
-                    }
-                    else if (subNode.stateMutability === 'payable') {
-                        stereotype = umlClass_1.OperatorStereotype.Payable;
-                    }
-                    umlClass.operators.push({
-                        visibility: parseVisibility(subNode.visibility),
-                        name: subNode.name,
-                        stereotype,
-                        parameters: parseParameters(subNode.parameters),
-                        returnParameters: parseParameters(subNode.returnParameters),
-                    });
-                }
-                // Recursively parse function parameters for associations
-                umlClass = addAssociations(subNode.parameters, umlClass);
-                if (subNode.returnParameters) {
-                    umlClass = addAssociations(subNode.returnParameters, umlClass);
-                }
-                // If no body to the function, it must be either an Interface or Abstract
+            }
+            // If a fallback function
+            else if (subNode.name === '') {
+                umlClass.operators.push({
+                    name: '',
+                    stereotype: umlClass_1.OperatorStereotype.Fallback,
+                    parameters: parseParameters(subNode.parameters),
+                    isPayable: parsePayable(subNode.stateMutability),
+                });
+            }
+            else {
+                let stereotype = umlClass_1.OperatorStereotype.None;
                 if (subNode.body === null) {
-                    if (umlClass.stereotype !== umlClass_1.ClassStereotype.Interface) {
-                        // If not Interface, it must be Abstract
-                        umlClass.stereotype = umlClass_1.ClassStereotype.Abstract;
-                    }
+                    stereotype = umlClass_1.OperatorStereotype.Abstract;
                 }
-                else {
-                    // Recursively parse function statements for associations
-                    umlClass = addAssociations(subNode.body.statements, umlClass);
+                else if (subNode.stateMutability === 'payable') {
+                    stereotype = umlClass_1.OperatorStereotype.Payable;
                 }
-                break;
-            case 'ModifierDefinition':
                 umlClass.operators.push({
-                    stereotype: umlClass_1.OperatorStereotype.Modifier,
+                    visibility: parseVisibility(subNode.visibility),
                     name: subNode.name,
+                    stereotype,
                     parameters: parseParameters(subNode.parameters),
+                    returnParameters: parseParameters(subNode.returnParameters),
                 });
-                if (subNode.body && subNode.body.statements) {
-                    // Recursively parse modifier statements for associations
-                    umlClass = addAssociations(subNode.body.statements, umlClass);
+            }
+            // Recursively parse function parameters for associations
+            umlClass = addAssociations(subNode.parameters, umlClass);
+            if (subNode.returnParameters) {
+                umlClass = addAssociations(subNode.returnParameters, umlClass);
+            }
+            // If no body to the function, it must be either an Interface or Abstract
+            if (subNode.body === null) {
+                if (umlClass.stereotype !== umlClass_1.ClassStereotype.Interface) {
+                    // If not Interface, it must be Abstract
+                    umlClass.stereotype = umlClass_1.ClassStereotype.Abstract;
                 }
-                break;
-            case 'EventDefinition':
-                umlClass.operators.push({
-                    stereotype: umlClass_1.OperatorStereotype.Event,
-                    name: subNode.name,
-                    parameters: parseParameters(subNode.parameters),
+            }
+            else {
+                // Recursively parse function statements for associations
+                umlClass = addAssociations(subNode.body.statements, umlClass);
+            }
+        }
+        else if (typeGuards_1.isModifierDefinition(subNode)) {
+            umlClass.operators.push({
+                stereotype: umlClass_1.OperatorStereotype.Modifier,
+                name: subNode.name,
+                parameters: parseParameters(subNode.parameters),
+            });
+            if (subNode.body && subNode.body.statements) {
+                // Recursively parse modifier statements for associations
+                umlClass = addAssociations(subNode.body.statements, umlClass);
+            }
+        }
+        else if (typeGuards_1.isEventDefinition(subNode)) {
+            umlClass.operators.push({
+                stereotype: umlClass_1.OperatorStereotype.Event,
+                name: subNode.name,
+                parameters: parseParameters(subNode.parameters),
+            });
+            // Recursively parse event parameters for associations
+            umlClass = addAssociations(subNode.parameters, umlClass);
+        }
+        else if (typeGuards_1.isStructDefinition(subNode)) {
+            let structMembers = [];
+            subNode.members.forEach((member) => {
+                structMembers.push({
+                    name: member.name,
+                    type: parseTypeName(member.typeName),
                 });
-                // Recursively parse event parameters for associations
-                umlClass = addAssociations(subNode.parameters, umlClass);
-                break;
-            case 'StructDefinition':
-                let structMembers = [];
-                subNode.members.forEach((member) => {
-                    structMembers.push({
-                        name: member.name,
-                        type: parseTypeName(member.typeName),
-                    });
-                });
-                umlClass.structs[subNode.name] = structMembers;
-                // Recursively parse members for associations
-                umlClass = addAssociations(subNode.members, umlClass);
-                break;
-            case 'EnumDefinition':
-                let enumValues = [];
-                subNode.members.forEach((member) => {
-                    enumValues.push(member.name);
-                });
-                umlClass.enums[subNode.name] = enumValues;
-                break;
+            });
+            umlClass.structs[subNode.name] = structMembers;
+            // Recursively parse members for associations
+            umlClass = addAssociations(subNode.members, umlClass);
+        }
+        else if (typeGuards_1.isEnumDefinition(subNode)) {
+            let enumValues = [];
+            subNode.members.forEach((member) => {
+                enumValues.push(member.name);
+            });
+            umlClass.enums[subNode.name] = enumValues;
         }
     });
     return umlClass;
@@ -364,6 +412,7 @@ function parseVisibility(visibility) {
     }
 }
 function parseTypeName(typeName) {
+    var _a, _b;
     switch (typeName.type) {
         case 'ElementaryTypeName':
             return typeName.name;
@@ -375,11 +424,10 @@ function parseTypeName(typeName) {
         case 'ArrayTypeName':
             return parseTypeName(typeName.baseTypeName) + '[]';
         case 'Mapping':
-            return ('mapping\\(' +
-                typeName.keyType.name +
-                '=\\>' +
-                parseTypeName(typeName.valueType) +
-                '\\)');
+            const key = ((_a = typeName.keyType) === null || _a === void 0 ? void 0 : _a.name) ||
+                ((_b = typeName.keyType) === null || _b === void 0 ? void 0 : _b.namePath);
+            const value = parseTypeName(typeName.valueType);
+            return 'mapping\\(' + key + '=\\>' + value + '\\)';
         default:
             throw Error(`Invalid typeName ${typeName}`);
     }
